requirements.txt
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-pubsub==2.6.1
google-cloud-pubsublite==1.0.0
googleapis-common-protos==1.53.0
google-cloud-storage==1.41.1
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
grpcio-status==1.39.0
toml==0.10.2
feedparser==6.0.8
